added documentation, added imgui widgets to support types other than int, cleaned up code, made godmode not depend on rings (by constantly checking ring count), added some features
