+++ SRB2cheat/Trampoline.h
@@ -13,7 +13,10 @@ typedef struct _HOOK_OBJECT {
 
 namespace Trampo {
 
+	//Activates the trampoline hook, takes pointer to hook object.
 	BOOL EnableHook(PHOOK_OBJECT HookObj);
+	//Writes back the stolen bytes, takes pointer to hook object.
 	BOOL DisableHook(PHOOK_OBJECT HookObj);
+	//Creates an hook object. takes hook pointer, pointer to entry of bytes to steal and the length to steal.
 	PHOOK_OBJECT CreateHook(LPVOID hook, LPVOID orig, size_t len);
 }
