+++ SRB2cheat/util.cpp
@@ -1,0 +1,17 @@
+#include "util.h"
+#include <stdio.h>
+
+void Util::getCon() {
+
+    AllocConsole();
+    AttachConsole(GetCurrentProcessId());
+    (void)freopen_s((FILE**)stdout, "CON", "w", stdout);
+    (void)freopen_s((FILE**)stdin, "CON", "r", stdin);
+
+}
+
+INT Util::getBitNum(BYTE byte) {
+    BYTE nibble_lookup[16] = { 0, 1, 1, 2, 1, 2, 2, 3, 1, 2, 2, 3, 2, 3, 3, 4 };
+
+    return nibble_lookup[byte & 0x0F] + nibble_lookup[byte >> 4];
+}
